The first tag created!  This is the old (by now) SDL OpenGL version of the engine.  Yes, we're still way past the AllegroGL version, but we're still old I'm sure.
Currently, we're in the process of updating the game engine to it's new structure.  SphereObjects, PolyObjects, DisplayMeshes, all those fun things in to the DynamicObject who contains a Body2D and Mesh2D.  Moving the code from the old "src" folder in to the now better "Code" folder, as well, the content out of the src folder in to it's own folder.  I'm asuming one day when we get lots more content, we're going to want to move content out another level, so the game projects aren't bogged down, but alas, we're not doing that yet.

Other fun trivia.
- We have a "rubber tree", also nicknamed the Jesus tree ('cause the graphic has a strong resemblance to that star of Bethlaham (sp?)).
- We have something called the "Cool Block", which is this sort of vampire looking purple block.  Really, this was our first cool looking object.
- We have cartoony body parts objects.  A stomach, which was a lot of fun to crash in to in our demonstrations.  We also had a heart, kidney, lungs, and brain.  These were part of a joke casual game mockup I did on April fools this year.  Har har.  Full body match!

Hooray!

git-svn-id: svn://10.0.50.55/puffbomb/tags/Tech@1210 d14347b8-48fd-0310-9e84-97907d434d85